Update RepeatedSweeps.py
Final Version. I made the plotting faster, and I rewrote the lakeshore interface so that will work for both model 325 and model 335. I also elminated one of the statuses. The status with the frequency sweep kept causing problems between the chair and keyboard. Streamlining the status sequencing will help to eliminate ways to screw up measurement